fix: implement proper BEEF ancestry collection in to_beef()
BREAKING: Transaction::to_beef() now correctly walks the sourceTransaction
chain to collect all ancestor transactions and merkle proofs, matching the
TypeScript and Go SDK behavior. Previously it only added the single target
transaction, making the BEEF invalid for SPV verification.

Changes:
- Add `merkle_path: Option<MerklePath>` field to Transaction struct
  This marks transactions as "proven" (mined with SPV proof) and tells
  the ancestry collection when to stop recursing

- Implement collect_ancestors() helper method that:
  - Uses depth-first traversal to walk input.source_transaction chains
  - Stops recursion at transactions with merkle_path (proven)
  - Uses HashSet for cycle detection and deduplication
  - Collects merkle proofs deduplicated by block height + root
  - Returns transactions in dependency order (ancestors first)
  - Respects the allow_partial flag

- Rewrite to_beef() and to_atomic_beef() to use ancestry collection

- Add Beef::make_txid_only() method to convert full transactions to
  txid-only format (for trimming known txs from returned BEEFs)

- Fix potential overflow in script.rs PUSHDATA comparison

- Add 13 comprehensive tests for ancestry collection covering:
  - Multi-level ancestry chains (2, 3, 4 levels)
  - Stopping at proven transactions
  - Merkle proof collection and deduplication
  - allow_partial flag behavior
  - Diamond dependency deduplication
  - Correct dependency ordering
  - Multiple input chains

- Add documentation tracking cross-SDK gaps and Go SDK parity

+++ src/script/script.rs
@@ -414,7 +414,7 @@ impl Script {
         } else if len < 65536 {
             op = OP_PUSHDATA2;
             data = Some(bin.to_vec());
-        } else if len < 0x100000000 {
+        } else if (len as u64) < 0x100000000 {
             op = OP_PUSHDATA4;
             data = Some(bin.to_vec());
         } else {
